vector2

+++ Vector/vector.h
@@ -6,12 +6,12 @@
 typedef struct int_vector
 {
     int *items;
-    int capacity;
-    int size;
+    size_t capacity;
+    size_t size;
 } int_vector;
 
 int_vector vector_init(size_t);
-int vector_size(int_vector *);
+size_t vector_size(int_vector *);
 bool vector_resize(int_vector *, size_t);
 void vector_push_back(int_vector *, int);
 bool vector_set(int_vector *, size_t, int);
